[refactor] user app rename
