+++ account/admin.py
@@ -4,7 +4,7 @@ from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 
 from .forms import UserCreationForm, UserChangeForm
 from .models import User
-import users.models
+import user.models
 
 
 class UserAdmin(BaseUserAdmin):
@@ -29,5 +29,5 @@ class UserAdmin(BaseUserAdmin):
     filter_horizontal = ()
 
 
-admin.site.register(User, UserAdmin, users.models.Group)
+admin.site.register(User, UserAdmin, user.models.Group)
 admin.site.unregister(Group)
